+++ apps/chat/src/components/Settings/SettingDialog.tsx
@@ -17,7 +17,6 @@ import { UIActions, UISelectors } from '@/src/store/ui/ui.reducers';
 import Modal from '@/src/components/Common/Modal';
 
 import { ToggleSwitchLabeled } from '../Common/ToggleSwitch/ToggleSwitchLabeled';
-import { FileManagerModal } from '../Files/FileManagerModal';
 import { CustomLogoSelect } from './CustomLogoSelect';
 import { ThemeSelect } from './ThemeSelect';
 
@@ -31,7 +30,6 @@ export const SettingDialog: FC<Props> = ({ open, onClose }) => {
   const isChatFullWidth = useAppSelector(UISelectors.selectIsChatFullWidth);
   const files = useAppSelector(FilesSelectors.selectFiles);
   const customLogoUrl = useAppSelector(UISelectors.selectCustomLogo);
-  const maximumAttachmentsAmount = 1;
   const customLogoId = customLogoUrl && ApiUtils.decodeApiUrl(customLogoUrl);
   const customLogoLocalStoreName = useMemo(() => {
     return customLogoId && splitEntityId(customLogoId).name;
@@ -47,9 +45,6 @@ export const SettingDialog: FC<Props> = ({ open, onClose }) => {
       }
     },
   );
-
-  const [isSelectFilesDialogOpened, setIsSelectFilesDialogOpened] =
-    useState(false);
 
   const saveBtnRef = useRef<HTMLButtonElement>(null);
 
@@ -79,18 +74,11 @@ export const SettingDialog: FC<Props> = ({ open, onClose }) => {
     setIsChatFullWidthLocal((prev) => !prev);
   }, []);
 
-  const onLogoSelect = (filesIds: string[]) => {
-    const selectedFileId = filesIds[0];
-    const newFile = files.find((file) => file.id === selectedFileId);
-    setLocalLogoFile(newFile);
-  };
-
   const handleSave = useCallback(() => {
     dispatch(UIActions.setTheme(localTheme));
     dispatch(UIActions.setIsChatFullWidth(isChatFullWidthLocal));
     if (localLogoFile) {
-      const logo = ApiUtils.encodeApiUrl(localLogoFile?.id);
-      dispatch(UIActions.setCustomLogo({ logo }));
+      dispatch(UIActions.setCustomLogo({ logo: localLogoFile?.id }));
     }
 
     onClose();
@@ -123,7 +111,8 @@ export const SettingDialog: FC<Props> = ({ open, onClose }) => {
           onThemeChangeHandler={onThemeChangeHandler}
         />
         <CustomLogoSelect
-          setOpenFilesModal={setIsSelectFilesDialogOpened}
+          files={files}
+          setLocalLogoFile={setLocalLogoFile}
           localLogo={
             (localLogoFile && localLogoFile.name) ?? customLogoLocalStoreName
           }
@@ -148,18 +137,6 @@ export const SettingDialog: FC<Props> = ({ open, onClose }) => {
           {t('Save')}
         </button>
       </div>
-      {isSelectFilesDialogOpened && (
-        <FileManagerModal
-          isOpen
-          allowedTypes={['image/*']}
-          maximumAttachmentsAmount={maximumAttachmentsAmount}
-          onClose={(files: unknown) => {
-            onLogoSelect(files as string[]);
-            setIsSelectFilesDialogOpened(false);
-          }}
-          isInLogoSelect
-        />
-      )}
     </Modal>
   );
 };
